requirements.txt
Django==2.0.7
django_filter==2.0.0
django_multiselectfield==0.1.8
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
coreapi
